backend/requirements.txt
--extra-index-url https://artifacts.dell.com:443/artifactory/api/pypi/agtsdk-1007569-pypi-prd-local/simple/
--trusted-host artifacts.dell.com
aia-auth-client
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.6.1
httpx==0.25.2
python-dotenv==1.0.0
apscheduler==3.10.4
jinja2==3.1.2
certifi
requests
